fix(mq_prometheus): correct first collection logic
The boolean logic for `isFirstCollection` was flipped in commit
`11efdcdcb9d576bd5c7a87a1f19925dba4e4ac71`. This caused
`!isFirstCollection()` to evaluate to false on the first poll, and true
after the first poll interval, leading to frequent status polling
instead of adhering to `global.pollInterval`.

This commit corrects the logic to ensure proper collection on the first
poll and at regular intervals thereafter.